+++ packages/throws_plugin/lib/main.dart
@@ -1,10 +1,13 @@
 import 'package:analysis_server_plugin/plugin.dart';
 import 'package:analysis_server_plugin/registry.dart';
 
+import 'src/assists/add_default_catch_clause_assist.dart';
+import 'src/assists/add_missing_catch_clauses_assist.dart';
 import 'src/assists/add_throws_annotation_assist.dart';
 import 'src/assists/remove_throws_annotation_assist.dart';
 import 'src/assists/update_throws_annotation_assist.dart';
 import 'src/assists/wrap_throws_call_in_try_catch_assist.dart';
+import 'src/assists/wrap_throws_call_in_try_catch_with_default_assist.dart';
 import 'src/lints/introduced_throws_in_override_rule.dart';
 import 'src/lints/missing_error_handling_rule.dart';
 import 'src/lints/missing_throws_annotation_rule.dart';
@@ -24,9 +27,12 @@ class ThrowsAnalyzerPlugin extends Plugin {
     registry.registerLintRule(ThrowsAnnotationMismatchRule());
     registry.registerLintRule(IntroducedThrowsInOverrideRule());
     registry.registerLintRule(UnusedThrowsAnnotationRule());
+    registry.registerAssist(AddDefaultCatchClauseAssist.new);
+    registry.registerAssist(AddMissingCatchClausesAssist.new);
     registry.registerAssist(AddThrowsAnnotationAssist.new);
     registry.registerAssist(RemoveThrowsAnnotationAssist.new);
     registry.registerAssist(WrapThrowsCallInTryCatchAssist.new);
+    registry.registerAssist(WrapThrowsCallInTryCatchWithDefaultAssist.new);
     registry.registerAssist(UpdateThrowsAnnotationAssist.new);
   }
 }