+++ packages/throws/CHANGELOG.md
@@ -1,4 +1,7 @@
-## 1.0.0-alpha.2
+## 1.0.0-alpha.4
+- Makes constructor named
+
+## 1.0.0-alpha.3
 - Fix README.md
 
 ## 1.0.0-alpha.2